DPC-678: Moved caveat tests back into macaroons module
Moved the caveats around, to avoid issues with test coverage not being reported correctly.
