+++ dpc-api/src/test/java/gov/cms/dpc/api/auth/macaroons/BakeryTests.java
@@ -4,12 +4,11 @@ import com.github.nitram509.jmacaroons.Macaroon;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigBeanFactory;
 import com.typesafe.config.ConfigFactory;
-import gov.cms.dpc.api.DPCAPIConfiguration;
-import gov.cms.dpc.macaroons.config.TokenPolicy;
 import gov.cms.dpc.macaroons.BakeryProvider;
 import gov.cms.dpc.macaroons.MacaroonBakery;
 import gov.cms.dpc.macaroons.MacaroonCaveat;
 import gov.cms.dpc.macaroons.MacaroonCondition;
+import gov.cms.dpc.macaroons.config.TokenPolicy;
 import gov.cms.dpc.macaroons.exceptions.BakeryException;
 import gov.cms.dpc.macaroons.store.MemoryRootKeyStore;
 import gov.cms.dpc.macaroons.thirdparty.MemoryThirdPartyKeyStore;
